store/wizard: consolidate filesystem types into submodule
Move PartitioningModeType, FilesystemMode, and FilesystemSlice from
the parent slice into the filesystem types module.
